修复 30 项确定性 bug: 全仓 UTF-16 代理对截断 + ANSI/JPEG/边框
- UTF-16 代理对截断清扫：Program/工具/记忆/会话约 23 处切片改用 TruncateByRunes/TruncateTailByRunes
- LSP：Content-Length 按字节、initialized 误读响应、ArgumentList 参数、file:// URI
- AnsiString：CSI 终止符按 0x40-0x7E 区间（含 DECTCEM 光标隐藏），跳过 ESC[ 引入符
- BoxBuffer：Width<2 负宽度崩溃 + 省略号 off-by-one
- TuiToastQueue/TuiDynamicBar 双省略号、InlinePermission/SessionPicker/DiffPreview 省略号预留宽度
- TuiRichEditor 宽度判据统一委托 CharWidth 真源
- JpegCodec DHT 表 4→16 防越界、--max-requeue 配置丢失、LLM Retry-After 整数溢出、ProjectContext .git 文件识别

+++ WayCoder/Agent/LLM.cs
@@ -743,7 +743,7 @@ public class LLM
 
             // 纯数字 = 秒数
             if (int.TryParse(header, out var seconds))
-                return Math.Min(seconds * 1000, Config.Instance.LlmRateLimitMaxWaitSec * 1000);
+                return (int)Math.Min((long)seconds * 1000, (long)Config.Instance.LlmRateLimitMaxWaitSec * 1000);
 
             // HTTP-date 格式
             if (DateTime.TryParse(header, out var retryDate))
